Beggining Categories
I made the frontend for the logic of the categories, and a little bit of the backend. I created a new table for the Categories and another table with an NxN relation to assign categories to notes. Im still missing the endpoins logic for creating a new Category. Adding categories to Notes comes after all that.